+++ CHANGELOG.md
@@ -6,6 +6,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [0.x.x] - Unreleased
 
+### Changed
+
+- using PyTorch's implementation of RAdam
+
 ## [0.2.0] - 2022/05/19
 
 ### Added